fix: resolve broken imports from moved/renamed modules (Phase 1)
Apply the import-path migrations catalogued in docs/dead-code-analysis.md
Phase 1. Each entry was either a context move (NotebookContext,
ThemeProvider), an editor-organism import that should resolve to the
Editor/overlays and Editor/extensions barrels, a missing barrel for
@/components/organisms/review, an IMetric fold into Metric, CDL
interpreter relative-path errors, or a runtime type rename
(IRuntimeBlockBuilder → BlockBuilder, IMemoryEntry → MemoryEntry, etc.).
The workbenchSyncStore store path entry is intentionally not in this
batch — that file is already on the @/ alias scheme and the doc's
relative-path prescription is stale.

Refs: docs/dead-code-analysis.md §1.1–1.6
