+++ src/components/organisms/review/interpreters/__tests__/cdlFilterInterpreter.test.ts
@@ -7,7 +7,7 @@ import {
 } from '../cdlFilterInterpreter';
 import { makeGridRow, makeGridCell } from './test-helpers';
 import { MetricType } from '@/core/models/Metric';
-import type { ColumnDef } from '../column-definition-language';
+import type { ColumnDef } from '../../column-definition-language';
 
 describe('cdlFilterInterpreter', () => {
   describe('extractFilterText', () => {
