+++ src/components/organisms/review/interpreters/__tests__/cdlGraphInterpreter.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'bun:test';
 import { extractGraphValue, buildGraphDataPoint } from '../cdlGraphInterpreter';
 import { makeGridRow, makeGridCell } from './test-helpers';
 import { MetricType } from '@/core/models/Metric';
-import type { ColumnDef } from '../column-definition-language';
+import type { ColumnDef } from '../../column-definition-language';
 
 describe('cdlGraphInterpreter', () => {
   describe('extractGraphValue', () => {
