+++ src/components/organisms/review/interpreters/__tests__/cdlSortInterpreter.test.ts
@@ -7,7 +7,7 @@ import {
 } from '../cdlSortInterpreter';
 import { makeGridRow, makeGridCell } from './test-helpers';
 import { MetricType } from '@/core/models/Metric';
-import type { ColumnDef } from '../column-definition-language';
+import type { ColumnDef } from '../../column-definition-language';
 
 describe('cdlSortInterpreter', () => {
   describe('extractSortValue', () => {
