+++ src/components/organisms/review/interpreters/__tests__/cdlSourceResolver.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'bun:test';
 import { resolveColumnSource } from '../cdlSourceResolver';
 import { makeGridRow, makeGridCell } from './test-helpers';
 import { MetricType } from '@/core/models/Metric';
-import type { ColumnSource } from '../column-definition-language';
+import type { ColumnSource } from '../../column-definition-language';
 
 describe('cdlSourceResolver', () => {
   describe('fixed-field source', () => {
